+++ R/f_core_plot.R
@@ -13,8 +13,8 @@
 ## |
 ## |  Contact us for information about our services: info@rpact.com
 ## |
-## |  File version: $Revision: 8619 $
-## |  Last changed: $Date: 2025-03-20 08:20:53 +0100 (Do, 20 Mrz 2025) $
+## |  File version: $Revision: 8629 $
+## |  Last changed: $Date: 2025-03-24 09:50:39 +0100 (Mo, 24 Mrz 2025) $
 ## |  Last changed by: $Author: pahlke $
 ## |
 
@@ -1763,4 +1763,32 @@ saveLastPlot <- function(filename, outputPath = .getRelativeFigureOutputPath()) 
     }
 
     return(.arrayToString(round(value, 2)))
+}
+
+.showWarningIfPlotArgumentWillBeIgnored <- function(type, ..., obj = NULL) {
+    if (any(type == "all") && (is.null(obj) || inherits(obj, "TrialDesignPlan"))) {
+        return(invisible())
+    }
+    
+    showFutilityBounds <- .getOptionalArgument("showFutilityBounds", ...)
+    if ((all(type != 3) || (!is.null(obj) && !inherits(obj, "TrialDesignPlan"))) && !is.null(showFutilityBounds)) {
+        objTypeInfo <- ifelse(!is.null(obj) && !inherits(obj, "TrialDesignPlan"), " design plan", "")
+        warning("Argument 'showFutilityBounds' (", showFutilityBounds, ") is only available for", objTypeInfo, " plot type 3; ",
+            "it will be ignored", call. = FALSE)
+    }
+    
+    showAlphaSpent <- .getOptionalArgument("showAlphaSpent", ...)
+    showBetaSpent <- .getOptionalArgument("showBetaSpent", ...)
+    if (all(type != 4) && all(type != "all") && (!is.null(showAlphaSpent) || !is.null(showBetaSpent))) {
+        if (!is.null(showAlphaSpent) && !is.null(showBetaSpent)) {
+            warning("Arguments 'showAlphaSpent' (", showAlphaSpent, ") and 'showBetaSpent' (", showBetaSpent, ") ",
+                "are only available for plot type 4; they will be ignored", call. = FALSE)
+        } else if (!is.null(showAlphaSpent)) {
+            warning("Argument 'showAlphaSpent' (", showAlphaSpent, ") is only available for plot type 4; ",
+                "it will be ignored", call. = FALSE)
+        } else {
+            warning("Argument 'showBetaSpent' (", showBetaSpent, ") is only available for plot type 4; ",
+                "it will be ignored", call. = FALSE)
+        }
+    }
 }