+++ NEWS.md
@@ -22,12 +22,14 @@
   Only members of the 'RPACT User Group' have access to the tests.
   For more information, please visit: [rpact.org/iq](https://www.rpact.org/vignettes/utilities/rpact_installation_qualification/) and [RPACT Connect](https://rpact.shinyapps.io/connect)
 * Usage of `maxInformation` improved (see enhancement [#65](https://github.com/rpact-com/rpact/issues/65))
+* Line breaks in the output of `getObjectRCode()` improved (see [#81](https://github.com/rpact-com/rpact/issues/81))
 * `testPackage()`: additional warning details will be added to the test report if warnings exist* Issue [#61](https://github.com/rpact-com/rpact/issues/61) fixed
 * Issue [#68](https://github.com/rpact-com/rpact/issues/68) fixed
 * Flexibility of function `getPiecewiseSurvivalTime()` improved
 * Simulation allows the case #events = #patients
 * Test coverage improved
 * Plot subtitles improved
+* Warning message added for extreme choice of `informationRates`, `userAlphaSpending`, and `userBetaSpending`  
 * Minor improvements